Background channel open

+++ app/src/main/java/to/bitkit/models/blocktank/BtOnchainTransaction.kt
@@ -4,12 +4,12 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class BtOnchainTransaction(
-    val amountSat: Int,
+    val amountSat: ULong,
     val txId: String,
     val vout: Int,
     val blockHeight: Int? = null,
     val blockConfirmationCount: Int,
-    val feeRateSatPerVbyte: Int,
+    val feeRateSatPerVbyte: Double,
     val confirmed: Boolean,
     val suspicious0ConfReason: String,
 )
